% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{max_package_list}
\alias{max_package_list}
\title{A full list of packages required}
\format{A long list of packages:
\describe{
  \item{gridExtra}
  \item{reshape2}
  \item{foreign}
  \item{ggrepel}
 ...
}}
\usage{
data(max_package_list)
}
\description{
A full list of packages required
}
\examples{
data(max_package_list)
}
\keyword{datasets}
