% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{min_package_list}
\alias{min_package_list}
\title{A minimum list of packages required}
\format{A list of length 4:
\describe{
  \item{tidyverse}
  \item{data.table}
  \item{feather}
  \item{knitr}
}}
\usage{
data(min_package_list)
}
\description{
A minimum list of packages required.
}
\examples{
data(min_package_list)
}
\keyword{datasets}
