% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{MyCheckSummary}
\alias{MyCheckSummary}
\title{Check basic summary statistics of a data frame}
\usage{
MyCheckSummary(DF)
}
\arguments{
\item{DF}{a data frame of interest}
}
\value{
a list that contains the number of rows, the number of columns, a data.table summarizing the number of unique values, the fraction of NAs, the class of each column, a data.table summarizing the mean, sd, min, max, fraction of NAs, number of unique values of each numeric columns.
}
\description{
Check basic summary statistics of a data frame
}
\examples{
MyCheckSummary(dt)

}
