% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot0_compute.R
\name{MyComputeMean}
\alias{MyComputeMean}
\title{Summarize col.to.y by mean}
\usage{
MyComputeMean(DF, col.to.y, col.to.groupby)
}
\arguments{
\item{DF}{data.table of interest}

\item{col.to.y}{the column with values to be summarized}

\item{col.to.groupby}{columns to group by}
}
\value{
a column where values are summarized by groups
}
\description{
Summarize col.to.y by mean
}
\examples{
MyComputeMean(dt, 'risk_score', 'race')

}
