% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot0_compute.R
\name{MyComputePercentile}
\alias{MyComputePercentile}
\title{Compute percentiles}
\usage{
MyComputePercentile(DF, col.to.cut, col.to.groupby)
}
\arguments{
\item{DF}{data.table of interest}

\item{col.to.cut}{the column with values to be ranked by percentiles}

\item{col.to.groupby}{columns to group by}
}
\value{
a column of percentile values
}
\description{
Compute percentiles
}
\examples{
MyComputePercentile(dt, 'risk_score', NULL)

}
