% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot0_compute.R
\name{MyComputeQuantile}
\alias{MyComputeQuantile}
\title{Compute quantiles}
\usage{
MyComputeQuantile(DF, nquantiles, col.to.cut, col.to.groupby)
}
\arguments{
\item{DF}{data.table of interest}

\item{nquantiles}{number of quantiles to be computed}

\item{col.to.cut}{the column with values to be ranked by percentiles}

\item{col.to.groupby}{columns to group by}
}
\value{
a column of quantile values
}
\description{
Compute quantiles
}
\examples{
MyComputeQuantile(dt, 10,'risk_score', NULL)

}
