% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot0_compute.R
\name{MyComputeSE}
\alias{MyComputeSE}
\title{Compute SE}
\usage{
MyComputeSE(DF, col.to.y, col.to.groupby)
}
\arguments{
\item{DF}{data.table of interest}

\item{col.to.y}{the column with values to be summarized}

\item{col.to.groupby}{columns to group by}
}
\value{
a column where values are summarized by groups
}
\description{
Compute SE
}
\examples{
MyComputeSE(dt, 'risk_score', 'race')

}
