% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{MyCreateDir}
\alias{MyCreateDir}
\title{Create a directory if not already exists}
\usage{
MyCreateDir(path = "")
}
\arguments{
\item{path}{a string of the path to be created}
}
\value{
create a directory at the path provided
}
\description{
Create a directory if not already exists
}
