% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{MyFread}
\alias{MyFread}
\title{Read text files with pre-specified path}
\usage{
MyFread(file, path, ...)
}
\arguments{
\item{path}{path to the data}

\item{fread}{the function that reads text files}

\item{to.data.table}{TRUE}
}
\value{
a data.table
}
\description{
Read text files with pre-specified path
}
