% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{MyGetVarName}
\alias{MyGetVarName}
\title{Get variable name as a string}
\usage{
MyGetVarName(var)
}
\arguments{
\item{var}{a variable}
}
\value{
name of the variable as a string
}
\description{
Get variable name as a string
}
