% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{MyInstall}
\alias{MyInstall}
\title{Install packages if not present}
\usage{
MyInstall(package_names)
}
\arguments{
\item{package_names}{a list of package to be installed}
}
\value{
install the list of packages provided.
}
\description{
Install packages if not present
}
\examples{
data(max_package_list)
MyInstall(max_package_list)

}
