% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{MyLoad}
\alias{MyLoad}
\title{Load necessary packages}
\usage{
MyLoad(package_names)
}
\arguments{
\item{package_names}{a list of packages to be loaded}
}
\value{
load the list of packages provided.
}
\description{
Load necessary packages
}
\examples{
data(min_package_list)
MyLoad(min_package_list)

}
