% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{MyReadCSV}
\alias{MyReadCSV}
\title{Read csv text files with pre-specified path}
\usage{
MyReadCSV(file, path, ...)
}
\arguments{
\item{path}{path to the data}

\item{read.csv}{the function that reads csv text files}

\item{to.data.table}{TRUE}
}
\value{
a data.table
}
\description{
Read csv text files with pre-specified path
}
