% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{MyReadFeather}
\alias{MyReadFeather}
\title{Read feather files with pre-specified path}
\usage{
MyReadFeather(file, path, ...)
}
\arguments{
\item{path}{path to the data}

\item{read_feather}{the function that reads feather files}

\item{to.data.table}{TRUE}
}
\value{
a data.table
}
\description{
Read feather files with pre-specified path
}
\examples{
feather_path <- '/data/zolab/racial_bias_final/data/master/new/'
enc_dt <- MyReadFeather('edw_enc_master_new_cohort.feather', path = feather_path)

}
