% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{MyReadRDS}
\alias{MyReadRDS}
\title{Read RDS files with pre-specified path}
\usage{
MyReadRDS(file, path, ...)
}
\arguments{
\item{path}{path to the data}

\item{readRDS}{the function that reads RDS files}

\item{to.data.table}{TRUE}
}
\value{
a data.table
}
\description{
Read RDS files with pre-specified path
}
\examples{
external_path <- '/external/SFA/PHMRES/research-fellow/Powers,\\ Brian/Jan\\ 2017\\ Merge/racial_bias_final/data/cohort/new/'
MyReadRDS('cohort_claims_biomarkers_bw.Rds', path = external_path)

}
