% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{MySink}
\alias{MySink}
\title{Print and sink output result to a text file}
\usage{
MySink(output_res, path, filename, width = "200")
}
\arguments{
\item{output_res}{an output to be printed and sinked}

\item{path}{path to file}

\item{filename}{name of the file}

\item{width}{width of the text file to be sinked}
}
\value{
print output result to a text file and export it to the path provided
}
\description{
Print and sink output result to a text file
}
\examples{
MySink(dt, "path/to/file", "filename", width = "200")

}
