% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{MySource}
\alias{MySource}
\title{Source files with pre-specified path}
\usage{
MySource(file, path, ...)
}
\arguments{
\item{path}{path to the data}

\item{source}{the function that source .R files}

\item{to.data.table}{FALSE}
}
\value{
objects from the sourced script
}
\description{
Source files with pre-specified path
}
\examples{
support_path <- '/data/zolab/racial_bias_final/code/support_code/'
MySource('exhibit_helper.R', path = support_path)

}
