% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pathit}
\alias{pathit}
\title{Generalize reading data and sourcing scripts}
\usage{
pathit(FUN, to.data.table = FALSE, path)
}
\arguments{
\item{FUN}{a function to read data}

\item{to.data.table}{whether to convert output result to a data.table}

\item{path}{path to the data}
}
\value{
a function with path preprogrammed in
}
\description{
Generalize reading data and sourcing scripts
}
\examples{
pathit(dt, "path/to/file", "filename", width = "200")

}
